+++ RELEASENOTES.md
@@ -1,5 +1,5 @@
-A Joomla 3 plugin to sort the articles in a category by the value of a custom field.
+A Joomla plugin to sort the articles in a category by the value of a custom field.
 
 **Requirements**
-* PHP 7.2 or later; or 8.0
-* Joomla 3.9; or 3.10
+* PHP 8.0, 8.1, or 8.2
+* Joomla 4.2 or later